[AIRFLOW-6680] Last changes for AIP-21 (#7301)
* Move postgres tests from test_operators to test_postgres

* Move myqsl tests from test_operators to test_mysql

* Move myqsl tests from test_operators to test_mysql

* Move myqsl_to_hive tests from test_operators

* Move GCP system tests to providers.google.cloud

* Move PythonVirtualenvOperator to test_python

* Move GenericTransfer tests to test_generic_transfer

* Move MSSQL tests to providers.microsoft.mssql.operators

* Remove duplicate test_to_gcs_to_s3

* [AIP-21] Move operators.presto_to_mysql operators.mysql.operators.presto_to_mysql

* Move Cassandra hook tests to test_cassandra

* Improve strcutre of cassandra senssor testss

* Move TestHiveCli to hooks.test_hive
