+++ tests/providers/apache/cassandra/hooks/test_cassandra.py
