+++ tests/providers/apache/hive/operators/test_mysql_to_hive.py
@@ -17,7 +17,6 @@
 # specific language governing permissions and limitations
 # under the License.
 
-import os
 import unittest
 from collections import OrderedDict
 from unittest import mock
@@ -25,247 +24,15 @@ from unittest import mock
 import pytest
 
 from airflow import DAG
+from airflow.providers.apache.hive.hooks.hive import HiveServer2Hook
+from airflow.providers.apache.hive.operators.mysql_to_hive import MySqlToHiveTransfer
+from airflow.providers.mysql.hooks.mysql import MySqlHook
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2015, 1, 1)
 DEFAULT_DATE_ISO = DEFAULT_DATE.isoformat()
 DEFAULT_DATE_DS = DEFAULT_DATE_ISO[:10]
 TEST_DAG_ID = 'unit_test_dag'
-
-
-@pytest.mark.backend("mysql")
-class TestMySql(unittest.TestCase):
-    def setUp(self):
-        args = {
-            'owner': 'airflow',
-            'start_date': DEFAULT_DATE
-        }
-        dag = DAG(TEST_DAG_ID, default_args=args)
-        self.dag = dag
-
-    def tearDown(self):
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
-        drop_tables = {'test_mysql_to_mysql', 'test_airflow'}
-        with MySqlHook().get_conn() as conn:
-            for table in drop_tables:
-                conn.execute("DROP TABLE IF EXISTS {}".format(table))
-
-    def test_mysql_operator_test(self):
-        sql = """
-        CREATE TABLE IF NOT EXISTS test_airflow (
-            dummy VARCHAR(50)
-        );
-        """
-        from airflow.providers.mysql.operators.mysql import MySqlOperator
-        op = MySqlOperator(
-            task_id='basic_mysql',
-            sql=sql,
-            dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_mysql_operator_test_multi(self):
-        sql = [
-            "CREATE TABLE IF NOT EXISTS test_airflow (dummy VARCHAR(50))",
-            "TRUNCATE TABLE test_airflow",
-            "INSERT INTO test_airflow VALUES ('X')",
-        ]
-        from airflow.providers.mysql.operators.mysql import MySqlOperator
-        op = MySqlOperator(
-            task_id='mysql_operator_test_multi',
-            sql=sql,
-            dag=self.dag,
-        )
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_mysql_hook_test_bulk_load(self):
-        records = ("foo", "bar", "baz")
-
-        import tempfile
-        with tempfile.NamedTemporaryFile() as f:
-            f.write("\n".join(records).encode('utf8'))
-            f.flush()
-
-            from airflow.providers.mysql.hooks.mysql import MySqlHook
-            hook = MySqlHook('airflow_db')
-            with hook.get_conn() as conn:
-                conn.execute("""
-                    CREATE TABLE IF NOT EXISTS test_airflow (
-                        dummy VARCHAR(50)
-                    )
-                """)
-                conn.execute("TRUNCATE TABLE test_airflow")
-                hook.bulk_load("test_airflow", f.name)
-                conn.execute("SELECT dummy FROM test_airflow")
-                results = tuple(result[0] for result in conn.fetchall())
-                self.assertEqual(sorted(results), sorted(records))
-
-    def test_mysql_hook_test_bulk_dump(self):
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
-        hook = MySqlHook('airflow_db')
-        priv = hook.get_first("SELECT @@global.secure_file_priv")
-        if priv and priv[0]:
-            # Confirm that no error occurs
-            hook.bulk_dump("INFORMATION_SCHEMA.TABLES", os.path.join(priv[0], "TABLES"))
-        else:
-            self.skipTest("Skip test_mysql_hook_test_bulk_load "
-                          "since file output is not permitted")
-
-    @mock.patch('airflow.providers.mysql.hooks.mysql.MySqlHook.get_conn')
-    def test_mysql_hook_test_bulk_dump_mock(self, mock_get_conn):
-        mock_execute = mock.MagicMock()
-        mock_get_conn.return_value.cursor.return_value.execute = mock_execute
-
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
-        hook = MySqlHook('airflow_db')
-        table = "INFORMATION_SCHEMA.TABLES"
-        tmp_file = "/path/to/output/file"
-        hook.bulk_dump(table, tmp_file)
-
-        from tests.test_utils.asserts import assert_equal_ignore_multiple_spaces
-        assert mock_execute.call_count == 1
-        query = """
-            SELECT * INTO OUTFILE '{tmp_file}'
-            FROM {table}
-        """.format(tmp_file=tmp_file, table=table)
-        assert_equal_ignore_multiple_spaces(self, mock_execute.call_args[0][0], query)
-
-    def test_mysql_to_mysql(self):
-        sql = "SELECT * FROM INFORMATION_SCHEMA.TABLES LIMIT 100;"
-        from airflow.operators.generic_transfer import GenericTransfer
-        op = GenericTransfer(
-            task_id='test_m2m',
-            preoperator=[
-                "DROP TABLE IF EXISTS test_mysql_to_mysql",
-                "CREATE TABLE IF NOT EXISTS "
-                "test_mysql_to_mysql LIKE INFORMATION_SCHEMA.TABLES"
-            ],
-            source_conn_id='airflow_db',
-            destination_conn_id='airflow_db',
-            destination_table="test_mysql_to_mysql",
-            sql=sql,
-            dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_overwrite_schema(self):
-        """
-        Verifies option to overwrite connection schema
-        """
-        from airflow.providers.mysql.operators.mysql import MySqlOperator
-
-        sql = "SELECT 1;"
-        op = MySqlOperator(
-            task_id='test_mysql_operator_test_schema_overwrite',
-            sql=sql,
-            dag=self.dag,
-            database="foobar",
-        )
-
-        from _mysql_exceptions import OperationalError
-        try:
-            op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE,
-                   ignore_ti_state=True)
-        except OperationalError as e:
-            assert "Unknown database 'foobar'" in str(e)
-
-
-@pytest.mark.backend("postgres")
-class TestPostgres(unittest.TestCase):
-    def setUp(self):
-        args = {'owner': 'airflow', 'start_date': DEFAULT_DATE}
-        dag = DAG(TEST_DAG_ID, default_args=args)
-        self.dag = dag
-
-    def tearDown(self):
-        tables_to_drop = ['test_postgres_to_postgres', 'test_airflow']
-        from airflow.providers.postgres.hooks.postgres import PostgresHook
-        with PostgresHook().get_conn() as conn:
-            with conn.cursor() as cur:
-                for table in tables_to_drop:
-                    cur.execute(f"DROP TABLE IF EXISTS {table}")
-
-    def test_postgres_operator_test(self):
-        sql = """
-        CREATE TABLE IF NOT EXISTS test_airflow (
-            dummy VARCHAR(50)
-        );
-        """
-        from airflow.providers.postgres.operators.postgres import PostgresOperator
-        op = PostgresOperator(task_id='basic_postgres', sql=sql, dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-        autocommit_task = PostgresOperator(
-            task_id='basic_postgres_with_autocommit',
-            sql=sql,
-            dag=self.dag,
-            autocommit=True)
-        autocommit_task.run(
-            start_date=DEFAULT_DATE,
-            end_date=DEFAULT_DATE,
-            ignore_ti_state=True)
-
-    def test_postgres_operator_test_multi(self):
-        sql = [
-            "CREATE TABLE IF NOT EXISTS test_airflow (dummy VARCHAR(50))",
-            "TRUNCATE TABLE test_airflow",
-            "INSERT INTO test_airflow VALUES ('X')",
-        ]
-        from airflow.providers.postgres.operators.postgres import PostgresOperator
-        op = PostgresOperator(
-            task_id='postgres_operator_test_multi', sql=sql, dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_postgres_to_postgres(self):
-        sql = "SELECT * FROM INFORMATION_SCHEMA.TABLES LIMIT 100;"
-        from airflow.operators.generic_transfer import GenericTransfer
-        op = GenericTransfer(
-            task_id='test_p2p',
-            preoperator=[
-                "DROP TABLE IF EXISTS test_postgres_to_postgres",
-                "CREATE TABLE IF NOT EXISTS "
-                "test_postgres_to_postgres (LIKE INFORMATION_SCHEMA.TABLES)"
-            ],
-            source_conn_id='postgres_default',
-            destination_conn_id='postgres_default',
-            destination_table="test_postgres_to_postgres",
-            sql=sql,
-            dag=self.dag)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_vacuum(self):
-        """
-        Verifies the VACUUM operation runs well with the PostgresOperator
-        """
-        from airflow.providers.postgres.operators.postgres import PostgresOperator
-
-        sql = "VACUUM ANALYZE;"
-        op = PostgresOperator(
-            task_id='postgres_operator_test_vacuum',
-            sql=sql,
-            dag=self.dag,
-            autocommit=True)
-        op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
-
-    def test_overwrite_schema(self):
-        """
-        Verifies option to overwrite connection schema
-        """
-        from airflow.providers.postgres.operators.postgres import PostgresOperator
-
-        sql = "SELECT 1;"
-        op = PostgresOperator(
-            task_id='postgres_operator_test_schema_overwrite',
-            sql=sql,
-            dag=self.dag,
-            autocommit=True,
-            database="foobar",
-        )
-
-        from psycopg2 import OperationalError
-        try:
-            op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE,
-                   ignore_ti_state=True)
-        except OperationalError as e:
-            assert 'database "foobar" does not exist' in str(e)
 
 
 @pytest.mark.backend("mysql")
@@ -331,7 +98,6 @@ class TestTransfer(unittest.TestCase):
             (1880, "Eugene", 0.00277, "boy"),
         ]
 
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
         with MySqlHook().get_conn() as cur:
             cur.execute('''
             CREATE TABLE IF NOT EXISTS baby_names (
@@ -346,17 +112,10 @@ class TestTransfer(unittest.TestCase):
             cur.execute("INSERT INTO baby_names VALUES(%s, %s, %s, %s);", row)
 
     def tearDown(self):
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
         with MySqlHook().get_conn() as cur:
             cur.execute("DROP TABLE IF EXISTS baby_names CASCADE;")
 
-    def test_clear(self):
-        self.dag.clear(
-            start_date=DEFAULT_DATE,
-            end_date=timezone.utcnow())
-
     def test_mysql_to_hive(self):
-        from airflow.providers.apache.hive.operators.mysql_to_hive import MySqlToHiveTransfer
         sql = "SELECT * FROM baby_names LIMIT 1000;"
         op = MySqlToHiveTransfer(
             task_id='test_m2h',
@@ -369,7 +128,6 @@ class TestTransfer(unittest.TestCase):
         op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
 
     def test_mysql_to_hive_partition(self):
-        from airflow.providers.apache.hive.operators.mysql_to_hive import MySqlToHiveTransfer
         sql = "SELECT * FROM baby_names LIMIT 1000;"
         op = MySqlToHiveTransfer(
             task_id='test_m2h',
@@ -384,7 +142,6 @@ class TestTransfer(unittest.TestCase):
         op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
 
     def test_mysql_to_hive_tblproperties(self):
-        from airflow.providers.apache.hive.operators.mysql_to_hive import MySqlToHiveTransfer
         sql = "SELECT * FROM baby_names LIMIT 1000;"
         op = MySqlToHiveTransfer(
             task_id='test_m2h',
@@ -401,7 +158,6 @@ class TestTransfer(unittest.TestCase):
     def test_mysql_to_hive_type_conversion(self, mock_load_file):
         mysql_table = 'test_mysql_to_hive'
 
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
         hook = MySqlHook()
 
         try:
@@ -418,7 +174,6 @@ class TestTransfer(unittest.TestCase):
                     )
                 """.format(mysql_table))
 
-            from airflow.providers.apache.hive.operators.mysql_to_hive import MySqlToHiveTransfer
             op = MySqlToHiveTransfer(
                 task_id='test_m2h',
                 hive_cli_conn_id='hive_cli_default',
@@ -444,7 +199,6 @@ class TestTransfer(unittest.TestCase):
         mysql_table = 'test_mysql_to_hive'
         hive_table = 'test_mysql_to_hive'
 
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
         hook = MySqlHook()
 
         try:
@@ -466,7 +220,6 @@ class TestTransfer(unittest.TestCase):
                     )
                 """.format(mysql_table, *db_record))
 
-            from airflow.providers.apache.hive.operators.mysql_to_hive import MySqlToHiveTransfer
             import unicodecsv as csv
             op = MySqlToHiveTransfer(
                 task_id='test_m2h',
@@ -481,7 +234,6 @@ class TestTransfer(unittest.TestCase):
                 dag=self.dag)
             op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
 
-            from airflow.providers.apache.hive.hooks.hive import HiveServer2Hook
             hive_hook = HiveServer2Hook()
             result = hive_hook.get_records("SELECT * FROM {}".format(hive_table))
             self.assertEqual(result[0], db_record)
@@ -493,7 +245,6 @@ class TestTransfer(unittest.TestCase):
         mysql_table = 'test_mysql_to_hive'
         hive_table = 'test_mysql_to_hive'
 
-        from airflow.providers.mysql.hooks.mysql import MySqlHook
         hook = MySqlHook()
 
         try:
@@ -532,7 +283,6 @@ class TestTransfer(unittest.TestCase):
                     )
                 """.format(mysql_table, *minmax))
 
-            from airflow.providers.apache.hive.operators.mysql_to_hive import MySqlToHiveTransfer
             op = MySqlToHiveTransfer(
                 task_id='test_m2h',
                 hive_cli_conn_id='hive_cli_default',
@@ -543,7 +293,6 @@ class TestTransfer(unittest.TestCase):
                 dag=self.dag)
             op.run(start_date=DEFAULT_DATE, end_date=DEFAULT_DATE, ignore_ti_state=True)
 
-            from airflow.providers.apache.hive.hooks.hive import HiveServer2Hook
             hive_hook = HiveServer2Hook()
             result = hive_hook.get_records("SELECT * FROM {}".format(hive_table))
             self.assertEqual(result[0], minmax)
