+++ tests/providers/google/cloud/operators/test_gcs_to_sftp_system_helper.py
