+++ tests/providers/microsoft/msqql/__init__.py
@@ -16,14 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.mysql.operators.presto_to_mysql`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.mysql.operators.presto_to_mysql import PrestoToMySqlTransfer  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.mysql.operators.presto_to_mysql`.",
-    DeprecationWarning, stacklevel=2
-)
