+++ tests/providers/microsoft/msqql/operators/test_mssql.py
@@ -21,8 +21,8 @@ import mock
 from parameterized import parameterized
 
 from airflow.models import Connection
-from airflow.operators.mssql_operator import MsSqlOperator
 from airflow.providers.microsoft.mssql.hooks.mssql import MsSqlHook
+from airflow.providers.microsoft.mssql.operators.mssql import MsSqlOperator
 from airflow.providers.odbc.hooks.odbc import OdbcHook
 
 ODBC_CONN = Connection(conn_id='test-odbc', conn_type='odbc', )
